Learning Sports Predictor App/requirements.txt
requests
psycopg2-binary
pybaseball
pandas